venn6 working fast (#18)
The test_venn6.c file is finding all solutions with 5,5,5,4,4,4 5-faces
in less than a second.
This is much faster than I expected, and I will drop all further
optimization work.
I believe this is generating 80 solutions twice through a mirror image. 

There seem to be no disconnected face cycle failures, so that check is
pushed to the very end